Separate DEVMODE gamestate
